+++ src/helpers/serialize-expression-for-debugging.ts
@@ -9,7 +9,7 @@ export function serializeExpressionForDebugging(expression: Expression): string 
   let nextSymbolIndex = 1;
   const knownSymbols: { [TSymbol in symbol]?: number } = {};
   return JSON.stringify(expression, (_, value) => {
-    if (typeof value === 'object' && value.constructor !== Object) {
+    if (typeof value === 'object' && value && value.constructor !== Object) {
       return {
         __type: value.constructor.name,
         ...value